Add management command to fill out run info from GiantBomb db
The command is cache_giantbomb_info. It assumes that the current run name,
with common 'category' strings filtered off is a decent enough query
string (fair warning: "mario 1" lists Super Mario Bros. for the NES
_last_).

Currently, it just normalizes the game name, sets the year of release, and
caches the GiantBomb db ID in case we ever need to grab more information
in the future (or want to create a separate `Game` table).