beginnings of a RESTful API to tracker objects
for the moment, this implements DRF for a couple of the obvious models
in tracker, to demonstrate the methodology and provide the stuff to
build off of with further details.

vanilla DRF for the most part, but to accommodate clients, the
implemented views do have an optional '?include=fieldname,fieldname'
parameter which pulls related serialized objects out of the tree and
makes them siblings to the requested object(s) in a dictionary. this is
done using a mixin against completed DRF stuff, so it's probably pretty
extendable/reusable