Remove the startrun and endrun fields from Prize
These fields were meant to provide a range of runs during which a prize
submitter would recommend the prize should be open for. They were
consequently used in filters for prizes and for start and end time
information when starttime and endtime were not provided.

They're being removed because they were not used very often.

[#163009909]