Allow limiting the max length of bid suggestions
When setting allowuseroptions on a parent bid, you can now also set
option_max_length, which controls the maximum number of characters that
bid suggestions for that parent bid may have. For example, if you set
option_max_length to 8, then suggestions may have at most 8 characters.
If users submit a suggestion that is longer than that, the suggestion is
rejected as invalid.

[#163161628]