Add CI via Azure Pipelines
This adds a basic pipeline for Azure Pipelines that runs the tracker's
tests using toplevel. AP's ubuntu image still doesn't have python 3.8 so
that's not included in the build matrix.

I've also changed our psycopg2 requirement to psycopg2-binary. This
simplifies installation for CI and new users, since static libraries
will not be required as we won't be building from source.