Add TypeScript, Babel preset, and convert entry files
This lets Babel continue working just like normal, while adding the option for typechecking using `yarn tsc` or `yarn tsc:watch`.

The reason for this approach instead of using `awesome-typescript-loader` or `ts-loader` is so that development can work without caring about type checking, but then types can be enforced in a pre-commit hook and on CI. Much more efficient!